JBQA-6006
complex JCA subsystem model
